Disable the NavBarFs before running this test
The Feature switches are stored in system properties, which
are held by the JVM, and so are carried across test
invocations. If this test is run as part of a suite, then a
previous test has turned the NavBarFS off and so this test
passes cleanly. To allow it to pass when run individually,
just ensure this condition is also true.

A future, further solution to this would be to provide the origin that is missing to allow the Nav Bar to work properly.