First pass at generating the Valid column from the other data
I've had to include the Customer Intent column as this is required to detect the cases where the customer selects a case that's not offered e.g. you can't select CY-1 when only CY and CY+1 are options.

This passes, but I think there is some detail missing. The Valid column seems to be a mash-up of many cases. We might be getting the right result but for the wrong reason because we're combining results and the overlap is correct but not expressing the right rules. Moreover, were missing some cases as we only have 288 rows where there should be 384. If we were to add these back in, it might show up the flawed logic.