Logs the name of the logger in shortened form
The logger is constructed by passing the Class of the place where it is used, then it outputs the abbreviated class name, e.g. 'c.ObligationsConnector', otherwise we'd get the name given at creation like 'application'.

There is one limitation with this in that we don't pick up the Logger config for 'application' from logback.xml anymore. I think the way Log4J tries to address this is to have a hierarchy of logger configurations around packages that can be used as fall-back if a specific one is not supplied. However, this is problematic for us in that our apps have a very flat package hierarchy with 'controllers' and 'views' etc at the top level. The app is not packaged under uk.gov.mtd.viewandchange or similar. This would mean we would have to either re-package the app, or add a logger config for every top-level package in logback.xml.