Initial version of OptOut V14
All issues found in V11 of the sheet have been addressed. The only outstanding point is scenario 105 (line 155 of the scala test file, line 108 on the v14 spreadsheet) where we are seeking clarification on the reason for this scenario being invalid (it is thought that if CY-1 is crystallised then a status will be generated for CY+1, so it cannot be blank, though it is thought that this scenario could occur correctly by the only business ceasing trading in CY and then a new business started in CY).