MISUV-7874 Refactor getNextUpdatesQuarterlyReportingContentChecks() to use the OptOutProposition
This will make it easier to unify it with nextUpdatesPageOptOutViewModel() as they will both be based on the OptOutProposition.

Note that the mocks have had to be moved to use the TaxYear rather than the integer for the end year. This embraces the domain type and addresses the primitive obsession smell around this.